Summary:
Due to potential perf issues with using same depthwise conv kernels for
perf channel depthwise conv, we opt for replicating the kernels and
adding per channel support to them.
Note that the large kernels files are largely duplication of original
kernels. Assembly kernels have little more modifications than intrinsics
ones.

Test Plan:
qnnpack tests.
q8dwconv-test
convolution-test

Differential Revision: [D21339042](https://our.internmc.facebook.com/intern/diff/D21339042)

[ghstack-poisoned]

+++ aten/src/ATen/native/quantized/cpu/qnnpack/src/fc-prepack.cc
@@ -20,7 +20,7 @@ PackBMatrix::PackBMatrix(
           "failed to create fully connected operator with requant scale of "
           "%.7g for output channel %d."
           "Scale must be finite and positive",
-          requantization_scales[i], i);
+          requantization_scales[i], (int)i);
       assert("QNNPACK Runtime Error.");
     }
   }
